Refactor StrictCamelBaseModel from base to core module
The `StrictCamelBaseModel` class has been moved from base.py to core.py to enhance modularity. All import statements across various files which were using this BaseModel subclass from the `base` module are now updated to import it from the `core` module.
